Fixing #84 - Block until termination of long-running processes
If the function we call returns:

- a promise, or
- a future, or
- a core.async channel

then it waits for that thing to be resolved or to have a
value, that will be the catual return  code. This prevents
shutdown for long-running processes.

This fix ONLY WORKS FOR THE JVM but it's a basis for a
similar change in the JS part.