Allow multiline values for :as in options via customized :summary-fn
The value for :as acts as an option's description. Like the :app and :command
:description, :as now optionally supports multiline text via a vector of
strings.

Notes:
Cli-matic takes advantage of clojure.tools.cli to format options for display
in help. Ideally clojure.tools.cli would accept multiline descriptions for
options, but it does not.

We take advantage of clojure.tools.cli's :summary-fn to override default
handling and support multilines.

This work includes the strengthening of the spec for descriptions. A description
must be a non-blank string - or - a collection of strings that includes at least
one non-blank string.

I also moved the beef of get-clojure-options under utils to localize
interactions with clojure.tools.cli

Fixes #94