Tweak description of :default in README
I was trying to add required switches (“required options”) to my CLI
and I kept searching this page for the word “require” and scanning the
docs, trying to figure out how to do this. For some reason I kept
missing the full contents of this bullet. I hope that these minor
tweaks might help make this behavior more findable in the future for
future folks (including perhaps future me).